formated some things, populated some of card js to add functionality to second page
