+++ views/cardinfo.handlebars
@@ -1,21 +1,11 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Movie Search</title>
-    <link rel="stylesheet" href="/css/card.css">
-    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBXoe4FUK0t3O6Ip1gLJaxYPF4f75WOud0&libraries=places"></script>
 
-    
-</head>
 
 <header>
   <nav>
     <ul class="navbar">
       <li class="navbar-item"><a href="#">Home</a></li>
       <li class="navbar-item active"><a href="#">Movies</a></li>
-      <li class="navbar-item"><a href="#">Sign Up</a></li>
+      <li class="navbar-item"><a href="/login">Log In</a></li>
     </ul>
   </nav>
 </header>
