+++ views/homepage.handlebars
@@ -4,7 +4,7 @@
       <ul class="navbar">
         <li class="navbar-item active"><a href="#">Home</a></li>
         <li class="navbar-item"><a href="#">Movies</a></li>
-        <li class="navbar-item"><a href="/login">Login In</a></li>
+        <li class="navbar-item"><a href="/login">Log In</a></li>
       </ul>
     </nav>
